Conexion de back al frontend la redireccion de roles

+++ src/main/java/com/example/restaurants/model/entity/lote.java
@@ -34,8 +34,8 @@ public class lote {
     private Date FechaVencimiento;
 
     @Column(name = "cantidadrecibido")
-    private Integer cantidadrecibido;
+    private Double cantidadrecibido;
 
     @Column(name = "cantidadactual")
-    private Integer cantidadactual;
+    private Double cantidadactual;
 }
